Apps can you build in App Inventor
The big advantage of App Inventor is the aforementioned instant gratification. If students have an Android phone they can put together a real app as a "hello world" exercise that they can have running on their phone.
Phenomenal tool for teaching understanding, basics, troubleshooting and persistence of programming. This is a great starter program for app building. If students have familiarity with Scratch they'll advance faster with MIT App Inventor.